requirements.txt
PyYAML==6.0.2
pypinyin==0.54.0